Refactor GameManager into mixins
